unify two LinearCombinations
